requirements.txt
google-genai==0.3.2
google-adk==0.1.0
pydantic==2.5.2
python-dotenv==1.0.0
typing-extensions==4.8.0 